# Conflicts:
#	zsxz-rcyh/src/main/java/com/hdsx/rcyh/controller/XccspzController.java

+++ zsxz-rcyh/src/main/java/com/hdsx/rcyh/service/impl/XccspzServiceImpl.java
@@ -21,8 +21,11 @@ public class XccspzServiceImpl implements XccspzService {
 
     @Override
     public int delete(String xccsId) {
-        System.out.println("11111111"+xccspzMapper.delete(xccsId));
-        return xccspzMapper.delete(xccsId);
+        if (xccspzMapper.delete(xccsId) > 0) {
+            System.out.println("fhg" + xccspzMapper.delete(xccsId));
+            return xccspzMapper.delete(xccsId);
+        }
+        return 0;
     }
 
 
@@ -34,7 +37,7 @@ public class XccspzServiceImpl implements XccspzService {
     }
 
 
-    @Override
+   /* @Override
     public int deletes(String[] xccsId) {
         int deleteNum = 0;
         for (int i = 0; i < xccsId.length; i++) {
@@ -42,10 +45,11 @@ public class XccspzServiceImpl implements XccspzService {
         }
         return deleteNum;
 
-    }
+    }*/
 
     @Override
     public int insert(Xccspz xccspz) {
+        System.out.println(xccspz);
         return xccspzMapper.insert(xccspz);
     }
